Bug fix for media type check
The album Inhumane Rampage by Dragonforce was recognized by
MusicBrainz, but the release media type was "Enhanced CD", so
did not pass the check for == "CD". So, modified this check
to see if the format (e.g., "CD") is in the media type.

Also adds more type hints to methods.