Updates some logic, bug fixes
Updates the logic in the main function that checks
udev information so that there are more guard clauses
to continue the while loop.

Updates logging to include more information and adds ability
to set logging level in the CLI.

Updates metadata formatting for tagging so that release/disc
level information is parsed once before iterating over tracks.
Also moves to using the .get() method on the dict so that does
not fail if key is missing.