Updates logging, fixes DiscID 'bug'
Logging improved to pring out the disc device (/dev) in each
log so know what is happening when ripping multiple discs.

Was bug where metadata class was not scanning correct drive.
This is fixed by requiring the /dev device to be passed on
class initialization, with the disc used when running the scan