Gets systemtray working
Is not the cleanest, but the system tray app is working similar to
autoMakeMKV. Need to go through and clean up some code.

Would like to add a Dialog that shows the results from the
MusicBrainz search for the first 5(?) results that may have the
same DiscID. This dialog will have a timeout where if not interact/
selection within 30 seconds, will choose the first/best option
and go with it. Otherwise, user can pick which disc to match to.

Also would like to add ability to upload DiscID through GUI for
missing ID. I believe I saw a tag for this in the musicbrainz API
response where can get link to submit ID. Click button takes you
to the website (like Picard) and then a dialog could appear that
says "Try inserting disc again after ID is submitted" and eject
disc.