Gets release selector timeout working
When a disc is inserted, if releases found for musicbrainz search,
then display a dialog showing the matching discs. This allows the
user to select which disc is the correct one to get metadata for.

This window has a 30 second timeout, wherein if the user does not
select something to do, the first result from the search will be
assumed as the correct one and used.

The user can choose to "Wait", disabling the timeout so that more
information can be gathered, "Ignore" the disc completely, or
"Save" (i.e., rip) the disc without waiting for the time out.

To do:
  Figure out if can add pop-up for disc id not exist on musicbrainz