Updates track title info
Realized that on some songs, titles are very long (e.g., some Rush
songs) and can contain forward slashs. This was causing conversion
to flac to fail as file names were invalid.

Have added an extra track to get all related works to each recording
on a release. From these related works, we get their title and check
if the full title contains the title of the work AND if the work title
is shorter than the full title. We do this for all returned works.
If there is a shorter title found, that is returned as the short
title, otherwise we just use the original title.

As this takes some time, the parsing of release information has
been moved to the ripper thread so that the GUI does not lock up.
This required a small change to the Ripper class call.