Bug fixes in udev and terminate
Was a bug in the udev watchdog where USB flash drives/HDD/etc. would
be scanned for disc ID. Got around this by check that the device has
the ID_CDROM property, which all optical drives should support, at least,
CD.

Also fixes a bug where in the dev device string was not passed to the
Ripper.terminate() method, causing program to crash, as is a required
argument.  Found this bug by ejecting a disc right after it had been
inserted.

Small change in coverart files where the disc number is now prepended
to the file name just like the tracks as different discs in a multi-
volume collection may have different coverart