Fixes disc number bug, adds label my media title
Was a bug were then total disc count was used for the disc number
for every disc in a collection.

Added check box to the release selection dialog to enable naming
by medium title. This is helpful if full albums are included in
a mulit-disc set, such as the Bonfire release from AC/DC. One may
want to label the "album" by the album name (e.g., Back in Black)
instead of having multiple discs under the Bonfire album.