Improves progress widget, more windows tweaks
The progress widget is now created before all track metadata is
grabbed from musicbrainz. This is done so that a notification
that metadata is being grabbed and rip will begin shortly can be
displayed on the progress widget. The label is removed when rip
starts.

Updates the get_vender() code to grab information on windows systems.