Settings updates and QThread fixes (#6)
* Updates user settings handling

New Settings() object for user settings. Object is initialized in
main __init__ file and used in the settings widget and ripper.py.

* Updates user settings handling (#4)

New Settings() object for user settings. Object is initialized in
main __init__ file and used in the settings widget and ripper.py.

* Bug fix for settings updates

Failed to update the quit() method for new Settings update

* Fix for QThread killing (#5)

Adds a wait() method to the DiscHandler class that checks if ripper
object is defined and waits on it to be finished. This is done
so that the finished/success/failure handlers can wait for any
thread(s) to fully finish before cleaning them up.